move stores to profilemodel
