+++ .vscode/settings.json
@@ -2,5 +2,8 @@
     "editor.codeActionsOnSave": {
         "source.fixAll.eslint": true
     },
-    "eslint.validate": ["javascript", "typescript"]
+    "eslint.validate": [
+        "javascript",
+        "typescript"
+    ]
 }
